requirements.txt
et_xmlfile==2.0.0
greenlet==3.2.0
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
psycopg2-binary==2.9.10
py4j==0.10.9.7
pyspark==3.5.5
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
tzdata==2025.2
